Move Pico dependency setup to utils
